+++ Packages/UGF.Module.Scenes/Runtime/ISceneLoadParameters.cs
@@ -1,0 +1,11 @@
+﻿using UnityEngine.SceneManagement;
+
+namespace UGF.Module.Scenes.Runtime
+{
+    public interface ISceneLoadParameters
+    {
+        LoadSceneMode AddMode { get; }
+        LocalPhysicsMode PhysicsMode { get; }
+        bool AllowActivation { get; }
+    }
+}