+++ Packages/UGF.Module.Scenes/Runtime/SceneInfoAssetBase.cs
@@ -1,8 +1,0 @@
-﻿using UGF.Builder.Runtime;
-
-namespace UGF.Module.Scenes.Runtime
-{
-    public abstract class SceneInfoAssetBase : BuilderAsset<ISceneInfo>
-    {
-    }
-}