+++ Packages/UGF.Module.Scenes/Runtime/SceneLoaderAsset.cs
@@ -1,0 +1,8 @@
+﻿using UGF.Builder.Runtime;
+
+namespace UGF.Module.Scenes.Runtime
+{
+    public abstract class SceneLoaderAsset : BuilderAsset<ISceneLoader>
+    {
+    }
+}