+++ Packages/UGF.Module.Scenes/Runtime/SceneLoaderAssetBase.cs
@@ -1,8 +1,0 @@
-﻿using UGF.Builder.Runtime;
-
-namespace UGF.Module.Scenes.Runtime
-{
-    public abstract class SceneLoaderAssetBase : BuilderAsset<ISceneLoader>
-    {
-    }
-}