+++ Packages/UGF.Module.Scenes/Runtime/SceneLoader`1.cs
@@ -1,0 +1,6 @@
+﻿namespace UGF.Module.Scenes.Runtime
+{
+    public abstract class SceneLoader<TInfo> : SceneLoader<TInfo, ISceneLoadParameters, ISceneUnloadParameters> where TInfo : class, ISceneInfo
+    {
+    }
+}