+++ Packages/UGF.Module.Scenes/Runtime/SceneModuleAsset.cs
@@ -1,7 +1,10 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using UGF.Application.Runtime;
 using UGF.EditorTools.Runtime.IMGUI.AssetReferences;
+using UGF.EditorTools.Runtime.IMGUI.Attributes;
 using UnityEngine;
+using UnityEngine.SceneManagement;
 
 namespace UGF.Module.Scenes.Runtime
 {
@@ -10,13 +13,13 @@ namespace UGF.Module.Scenes.Runtime
     {
         [SerializeField] private bool m_unloadTrackedScenesOnUninitialize = true;
         [SerializeField] private bool m_registerApplicationForScenes = true;
-        [SerializeField] private List<AssetReference<SceneLoaderAssetBase>> m_loaders = new List<AssetReference<SceneLoaderAssetBase>>();
-        [SerializeField] private List<AssetReference<SceneInfoAssetBase>> m_scenes = new List<AssetReference<SceneInfoAssetBase>>();
+        [SerializeField] private List<AssetReference<SceneLoaderAsset>> m_loaders = new List<AssetReference<SceneLoaderAsset>>();
+        [SerializeField] private List<AssetReference<SceneGroupAsset>> m_groups = new List<AssetReference<SceneGroupAsset>>();
 
         public bool UnloadTrackedScenesOnUninitialize { get { return m_unloadTrackedScenesOnUninitialize; } set { m_unloadTrackedScenesOnUninitialize = value; } }
         public bool RegisterApplicationForScenes { get { return m_registerApplicationForScenes; } set { m_registerApplicationForScenes = value; } }
-        public List<AssetReference<SceneLoaderAssetBase>> Loaders { get { return m_loaders; } }
-        public List<AssetReference<SceneInfoAssetBase>> Scenes { get { return m_scenes; } }
+        public List<AssetReference<SceneLoaderAsset>> Loaders { get { return m_loaders; } }
+        public List<AssetReference<SceneGroupAsset>> Groups { get { return m_groups; } }
 
         protected override IApplicationModuleDescription OnBuildDescription()
         {
@@ -29,18 +32,17 @@ namespace UGF.Module.Scenes.Runtime
 
             for (int i = 0; i < m_loaders.Count; i++)
             {
-                AssetReference<SceneLoaderAssetBase> reference = m_loaders[i];
+                AssetReference<SceneLoaderAsset> reference = m_loaders[i];
                 ISceneLoader loader = reference.Asset.Build();
 
                 description.Loaders.Add(reference.Guid, loader);
             }
 
-            for (int i = 0; i < m_scenes.Count; i++)
+            for (int i = 0; i < m_groups.Count; i++)
             {
-                AssetReference<SceneInfoAssetBase> reference = m_scenes[i];
-                ISceneInfo scene = reference.Asset.Build();
+                SceneGroupAsset group = m_groups[i].Asset;
 
-                description.Scenes.Add(reference.Guid, scene);
+                group.GetScenes(description.Scenes);
             }
 
             return description;