+++ Packages/UGF.Module.Scenes/Runtime/SceneUnloadHandler.cs
@@ -2,5 +2,5 @@
 
 namespace UGF.Module.Scenes.Runtime
 {
-    public delegate void SceneUnloadHandler(string id, Scene scene, SceneUnloadParameters parameters);
+    public delegate void SceneUnloadHandler(string id, Scene scene, ISceneUnloadParameters parameters);
 }